refactor(vfs): separe the logic of file and vfs operations
