+++ Include/Kernel/FileSystem/VFS/File/FileOperations.h
@@ -10,17 +10,6 @@ namespace FileSystem {
 enum class VNodeType : LibC::uint8_t;
 struct VNode;
 
-struct FileStat {
-    LibC::uint64_t size;
-    LibC::uint32_t permissions;
-    LibC::uint64_t inode;
-
-    LibC::uint32_t uid;
-    LibC::uint32_t gid;
-
-    VNodeType type;
-};
-
 struct FileHandle {
     VNode* vnode;
     LibC::uint64_t offset;
