+++ Include/Kernel/FileSystem/VFS/VNodeType.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <LibC/stdint.h>
+
+namespace FileSystem {
+
+enum class VNodeType : LibC::uint8_t {
+    Unknown,
+    File,
+    Directory,
+    Symlink,
+    Device,
+    Mountpoint
+};
+}
